New app Mirror Sequence
Doesn't work. A lot of problems seem to come from not using the base
singleton for ArcadeButtons. I tried redifing another with new
callbacks, deleting the old one... I'm getting inconsistent and hard to
interpret behavior. Reverting to adding a released_flag to
ArcadeButtons, and not changing the singleton for this game.